Add files via upload
Servo_Prediction_using_Linear_Regression_project

A Servo Prediction model, also known as a servo control or servo system, is a control system that uses feedback to accurately position or control the motion of a mechanical device, such as a motor or an actuator. The goal of a servo system is to maintain a desired position or trajectory by continuously monitoring the actual position and making adjustments as needed.

In Python, you can develop a servo prediction model using various libraries and techniques.